:recycle: (emailSchema) move from features to shared/form
